pack/unpack: ensure 64bit size_t variables
On Windows long is 32 bits and extending to 64 bits is implementation
defined.

Make all variables used in 64bit pack calculations explicitly size_t.

+++ builtin/unpack-objects.c
@@ -440,9 +440,9 @@ static void unpack_delta_entry(enum object_type type, size_t delta_size,
 
 static void unpack_one(unsigned nr)
 {
-	unsigned shift;
+	size_t shift;
 	unsigned char *pack;
-	unsigned long c;
+	size_t c;
 	size_t size;
 	enum object_type type;
